+++ src/Components/src/Page/ArcFileEditor/ArcFileFooterTabs.fs
@@ -214,7 +214,8 @@ type ArcFileFooterTabs =
         (
             elementRef: IRefValue<option<Browser.Types.HTMLElement>>,
             setEditorMode: int option -> unit,
-            deleteTable: int -> unit
+            deleteTable: int -> unit,
+            onDeleteDataMap: unit -> unit
         ) =
 
 
@@ -244,6 +245,16 @@ type ArcFileFooterTabs =
                                 prop.className "swt:iconify swt:fluent--delete-20-filled swt:size-4"
                             ],
                         onClick = delete index
+                    )
+                  ]
+                | Some ActiveView.DataMap -> [
+                    ContextMenuItem(
+                        Html.span "Delete DataMap",
+                        icon =
+                            Html.i [
+                                prop.className "swt:iconify swt:fluent--delete-20-filled swt:size-4"
+                            ],
+                        onClick = (fun _ -> onDeleteDataMap ())
                     )
                   ]
                 | _ -> []
@@ -282,8 +293,13 @@ type ArcFileFooterTabs =
 
     [<ReactComponent>]
     static member Main
-        (arcFile: ArcFiles, activeView: ActiveView, setActiveView: ActiveView -> unit, setArcFile: ArcFiles -> unit)
-        =
+        (
+            arcFile: ArcFiles,
+            activeView: ActiveView,
+            setActiveView: ActiveView -> unit,
+            setArcFile: ArcFiles -> unit,
+            onDeleteDataMap: unit -> unit
+        ) =
         let tables = arcFile.ArcTables()
         let canAddTable = arcFile.CanCreateTables()
         let canRenderDataMap = arcFile.CanRenderDataMapView()
@@ -439,7 +455,7 @@ type ArcFileFooterTabs =
             )
 
         React.Fragment [
-            ArcFileFooterTabs.ContextMenu(tabsRef, setEditorMode, deleteTable)
+            ArcFileFooterTabs.ContextMenu(tabsRef, setEditorMode, deleteTable, onDeleteDataMap)
             Html.div [
                 prop.className "swt:bg-base-300"
                 prop.children [